requirements.txt
streamlit
easyocr
pillow
numpy
pandas
gspread
oauth2client
torch